requirements.txt
fastapi
uvicorn
sentence-transformers
requests
streamlit
python-dotenv
